[Xtend] support standalone compilation of Xtend files
https://bugs.eclipse.org/bugs/show_bug.cgi?id=342759

+++ plugins/org.eclipse.xtext.xtend2.standalone/src/org/eclipse/xtext/xtend2/compiler/batch/Xtend2BatchCompiler.java
@@ -7,7 +7,6 @@ import static org.eclipse.xtext.util.Strings.*;
 
 import java.io.File;
 import java.io.FileFilter;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.Writer;
@@ -18,10 +17,10 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.common.util.WrappedException;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
 import org.eclipse.jdt.core.compiler.batch.BatchCompiler;
-import org.eclipse.xtend2.lib.StringConcatenation;
 import org.eclipse.xtext.common.types.JvmGenericType;
 import org.eclipse.xtext.common.types.access.impl.ClasspathTypeProvider;
 import org.eclipse.xtext.common.types.access.impl.IndexedJvmTypeAccess;
@@ -38,7 +37,6 @@ import org.eclipse.xtext.resource.IResourceServiceProvider;
 import org.eclipse.xtext.resource.XtextResource;
 import org.eclipse.xtext.resource.XtextResourceSet;
 import org.eclipse.xtext.resource.impl.ResourceSetBasedResourceDescriptions;
-import org.eclipse.xtext.util.Files;
 import org.eclipse.xtext.util.Strings;
 import org.eclipse.xtext.validation.CheckMode;
 import org.eclipse.xtext.validation.IResourceValidator;
@@ -60,7 +58,14 @@ import com.google.inject.Provider;
  * @author Michael Clay - Initial contribution and API
  */
 public class Xtend2BatchCompiler {
-	protected final static Logger log = Logger.getLogger(Xtend2BatchCompiler.class.getName());
+	private final static Logger log = Logger.getLogger(Xtend2BatchCompiler.class.getName());
+
+	private static final FileFilter ACCEPT_ALL_FILTER = new FileFilter() {
+		public boolean accept(File pathname) {
+			return true;
+		}
+	};
+
 	@Inject
 	private Provider<ResourceSet> resourceSetProvider;
 	@Inject
@@ -78,12 +83,86 @@ public class Xtend2BatchCompiler {
 	@Inject
 	private IndexedJvmTypeAccess indexedJvmTypeAccess;
 
+	private Writer outputWriter;
+	private Writer errorWriter;
 	private String sourcePath;
 	private String classPath;
 	private String outputPath;
 	private String complianceLevel = "1.5";
-	private boolean verbose = true;
+	private boolean verbose = false;
+	private String tempDirectory = System.getProperty("java.io.tmpdir");
+	private boolean deleteTempDirectory = true;
 	private List<File> tempFolders = Lists.newArrayList();
+
+	public String getTempDirectory() {
+		return tempDirectory;
+	}
+
+	public void setTempDirectory(String tempDirectory) {
+		this.tempDirectory = tempDirectory;
+	}
+
+	public boolean isDeleteTempDirectory() {
+		return deleteTempDirectory;
+	}
+
+	public void setDeleteTempDirectory(boolean deletetempDirectory) {
+		this.deleteTempDirectory = deletetempDirectory;
+	}
+
+	public Writer getOutputWriter() {
+		if (outputWriter == null) {
+			outputWriter = new Writer() {
+				@Override
+				public void write(char[] data, int offset, int count) throws IOException {
+					String message = String.copyValueOf(data, offset, count);
+					if (!Strings.isEmpty(message.trim())) {
+						log.info(message);
+					}
+				}
+
+				@Override
+				public void flush() throws IOException {
+				}
+
+				@Override
+				public void close() throws IOException {
+				}
+			};
+		}
+		return outputWriter;
+	}
+
+	public void setOutputWriter(Writer ouputWriter) {
+		this.outputWriter = ouputWriter;
+	}
+
+	public Writer getErrorWriter() {
+		if (errorWriter == null) {
+			errorWriter = new Writer() {
+				@Override
+				public void write(char[] data, int offset, int count) throws IOException {
+					String message = String.copyValueOf(data, offset, count);
+					if (!Strings.isEmpty(message.trim())) {
+						log.error(message);
+					}
+				}
+
+				@Override
+				public void flush() throws IOException {
+				}
+
+				@Override
+				public void close() throws IOException {
+				}
+			};
+		}
+		return errorWriter;
+	}
+
+	public void setErrorWriter(Writer errorWriter) {
+		this.errorWriter = errorWriter;
+	}
 
 	public void setClassPath(String classPath) {
 		this.classPath = classPath;
@@ -101,24 +180,34 @@ public class Xtend2BatchCompiler {
 		return complianceLevel;
 	}
 
+	public void setVerbose(boolean verbose) {
+		this.verbose = verbose;
+	}
+
 	private boolean isVerbose() {
 		return verbose;
 	}
 
-	public void compile() {
+	public boolean compile() {
 		try {
 			ResourceSet resourceSet = loadXtendFiles();
-			File sourceDirectory = createClassStubs(resourceSet);
-			File classDirectory = preCompile(sourceDirectory);
+			File sourceDirectory = createStubs(resourceSet);
+			File classDirectory = createTempDir("classes");
+			if (!preCompileStubs(sourceDirectory, classDirectory)) {
+				return false;
+			}
 			installJvmTypeProvider(resourceSet, classDirectory);
 			List<Issue> issues = validate(resourceSet);
 			if (reportIssues(issues)) {
-				return;
+				return false;
 			}
 			generateJavaFiles(resourceSet);
 		} finally {
-			deleteTmpFolders();
+			if (isDeleteTempDirectory()) {
+				deleteTmpFolders();
+			}
 		}
+		return true;
 	}
 
 	protected ResourceSet loadXtendFiles() {
@@ -130,8 +219,8 @@ public class Xtend2BatchCompiler {
 			public boolean apply(URI input) {
 				boolean matches = nameBasedFilter.matches(input);
 				if (matches) {
-					if (log.isInfoEnabled()) {
-						log.info("Load '" + input + "'");
+					if (log.isDebugEnabled()) {
+						log.debug("load xtend file '" + input + "'");
 					}
 					resourceSet.getResource(input, true);
 				}
@@ -141,8 +230,8 @@ public class Xtend2BatchCompiler {
 		return resourceSet;
 	}
 
-	private File createClassStubs(ResourceSet resourceSet) {
-		File outputDirectory = createTempDir();
+	private File createStubs(ResourceSet resourceSet) {
+		File outputDirectory = createTempDir("stubs");
 		JavaIoFileSystemAccess fileSystemAccess = javaIoFileSystemAccessProvider.get();
 		fileSystemAccess.setOutputPath(outputDirectory.toString());
 		for (Resource resource : resourceSet.getResources()) {
@@ -150,24 +239,26 @@ public class Xtend2BatchCompiler {
 			if (xtendClass == null) {
 				continue;
 			}
-			StringConcatenation classSignatureBuilder = new StringConcatenation();
+			StringBuilder classSignatureBuilder = new StringBuilder();
 			classSignatureBuilder.append("package " + xtendClass.getPackageName() + ";");
-			classSignatureBuilder.newLine();
+			classSignatureBuilder.append("\n");
 			classSignatureBuilder.append("public class " + xtendClass.getName() + "{}");
+			if (log.isDebugEnabled()) {
+				log.debug("create java stub '" + getJavaFileName(xtendClass) + "'");
+			}
 			fileSystemAccess.generateFile(getJavaFileName(xtendClass), classSignatureBuilder.toString());
 		}
 		return outputDirectory;
 	}
 
-	private File preCompile(File tmpSourceDirectory) {
-		File classDirectory = createTempDir();
+	private boolean preCompileStubs(File tmpSourceDirectory, File classDirectory) {
 		List<String> commandLine = Lists.newArrayList();
 		// todo args
 		if (isVerbose()) {
 			commandLine.add("-verbose");
 		}
 		if (!isEmpty(classPath)) {
-			commandLine.add("-cp " + concat(" ", getClassPathEntries()));
+			commandLine.add("-cp " + concat(File.pathSeparator, getClassPathEntries()));
 		}
 		commandLine.add("-d " + classDirectory.toString());
 		commandLine.add("-" + getComplianceLevel());
@@ -175,21 +266,11 @@ public class Xtend2BatchCompiler {
 		List<String> sourceDirectories = newArrayList(getSourcePathDirectories());
 		sourceDirectories.add(tmpSourceDirectory.toString());
 		commandLine.add(concat(" ", sourceDirectories));
-		BatchCompiler.compile(concat(" ", commandLine), getPrintWriter(), getPrintWriter(), null);
-		return classDirectory;
-	}
-
-	protected PrintWriter getPrintWriter() {
-		return new PrintWriter(new Writer() {
-			@Override
-			public void write(char[] cbuf, int off, int len) throws IOException {
-				log.debug(String.copyValueOf(cbuf, off, len));
-			}
-			@Override
-			public void flush() throws IOException {}
-			@Override
-			public void close() throws IOException {}
-		});
+		if (log.isDebugEnabled()) {
+			log.debug("invoke batch compiler with '" + concat(" ", commandLine) + "'");
+		}
+		return BatchCompiler.compile(concat(" ", commandLine), new PrintWriter(getOutputWriter()), new PrintWriter(
+				getErrorWriter()), null);
 	}
 
 	private List<Issue> validate(ResourceSet resourceSet) {
@@ -224,9 +305,10 @@ public class Xtend2BatchCompiler {
 			}
 		});
 		if (log.isDebugEnabled()) {
-			log.debug("Classpath used for Xtend compilation : "+classPathUrls);
+			log.debug("classpath used for Xtend compilation : " + classPathUrls);
 		}
-		URLClassLoader urlClassLoader = new URLClassLoader(toArray(classPathUrls, URL.class), getClass().getClassLoader());
+		URLClassLoader urlClassLoader = new URLClassLoader(toArray(classPathUrls, URL.class), getClass()
+				.getClassLoader());
 		new ClasspathTypeProvider(urlClassLoader, resourceSet, indexedJvmTypeAccess);
 		((XtextResourceSet) resourceSet).setClasspathURIContext(urlClassLoader);
 	}
@@ -243,7 +325,7 @@ public class Xtend2BatchCompiler {
 					issueBuilder.append(resourceUri.toFileString());
 				}
 				issueBuilder.append("\n").append(issue.getLineNumber()).append(": ").append(issue.getMessage());
-				log.error(issueBuilder);
+				log.error(issueBuilder.toString());
 				hasErrorOrWarnings = true;
 			}
 		}
@@ -254,13 +336,22 @@ public class Xtend2BatchCompiler {
 		JavaIoFileSystemAccess javaIoFileSystemAccess = javaIoFileSystemAccessProvider.get();
 		javaIoFileSystemAccess.setOutputPath(outputPath);
 		ResourceSetBasedResourceDescriptions resourceDescriptions = getResourceDescriptions(resourceSet);
-		for (IEObjectDescription eObjectDescription : resourceDescriptions
-				.getExportedObjectsByType(Xtend2Package.eINSTANCE.getXtendClass())) {
+		Iterable<IEObjectDescription> exportedObjectsByType = resourceDescriptions
+				.getExportedObjectsByType(Xtend2Package.eINSTANCE.getXtendClass());
+		if (log.isInfoEnabled()) {
+			int size = Iterables.size(exportedObjectsByType);
+			if (size == 0) {
+				log.info("No sources to compile in '" + sourcePath + "'");
+			} else {
+				log.info("Compiling " + size + " source " + (size == 1 ? "file" : "files") + " to " + outputPath);
+			}
+		}
+		for (IEObjectDescription eObjectDescription : exportedObjectsByType) {
 			XtendClass xtendClass = (XtendClass) eObjectDescription.getEObjectOrProxy();
 			JvmGenericType jvmGenericType = xtend2JvmAssociations.getInferredType(xtendClass);
 			CharSequence generatedType = generator.generateType(jvmGenericType);
-			if (log.isInfoEnabled()) {
-				log.info("Generate '" + getJavaFileName(xtendClass) + "'");
+			if (log.isDebugEnabled()) {
+				log.debug("write '" + outputPath + File.separator + getJavaFileName(xtendClass) + "'");
 			}
 			javaIoFileSystemAccess.generateFile(getJavaFileName(xtendClass), generatedType);
 		}
@@ -300,42 +391,67 @@ public class Xtend2BatchCompiler {
 	}
 
 	protected List<String> getDirectories(String path) {
-		if (Strings.isEmpty(path))
-			return new java.util.ArrayList<String>();
-		final List<String> split = split(emptyIfNull(path), getPathSeparator());
-		return transform(split, new Function<String,String>(){
+		if (Strings.isEmpty(path)) {
+			return Lists.newArrayList();
+		}
+		final List<String> split = split(emptyIfNull(path), File.pathSeparator);
+		return transform(split, new Function<String, String>() {
 			public String apply(String from) {
 				try {
 					return new File(from).getCanonicalPath();
 				} catch (IOException e) {
-					throw new RuntimeException(e);
+					throw new WrappedException(e);
 				}
-			}});
+			}
+		});
 	}
 
-	protected String getPathSeparator() {
-		return System.getProperty("path.separator");
-	}
-
-	protected File createTempDir() {
-		File baseDir = new File(System.getProperty("java.io.tmpdir"));
-		File tempDir = new File(baseDir, "tmpdir" + String.valueOf(System.nanoTime()));
-		tempDir.mkdir();
+	private File createTempDir(String prefix) {
+		File tempDir = new File(getTempDirectory(), prefix);
+		cleanFolder(tempDir, ACCEPT_ALL_FILTER, true, true);
+		if (!tempDir.mkdirs()) {
+			throw new RuntimeException("Error creating temp directory '" + tempDir.getAbsolutePath() + "'");
+		}
 		tempFolders.add(tempDir);
 		return tempDir;
 	}
 
-	protected void deleteTmpFolders() {
+	private void deleteTmpFolders() {
 		for (File file : tempFolders) {
-			try {
-				Files.cleanFolder(file, new FileFilter() {
-					public boolean accept(File pathname) {
-						return true;
-					}
-				}, true, true);
-			} catch (FileNotFoundException fileNotFoundException) {
+			cleanFolder(file, ACCEPT_ALL_FILTER, true, true);
+		}
+	}
+
+	// FIXME: use Files#cleanFolder after the maven distro availability of version 2.2.x
+	private static boolean cleanFolder(File parentFolder, FileFilter filter, boolean continueOnError,
+			boolean deleteParentFolder) {
+		if (!parentFolder.exists()) {
+			return true;
+		}
+		if (filter == null)
+			filter = ACCEPT_ALL_FILTER;
+		log.debug("Cleaning folder " + parentFolder.toString());
+		final File[] contents = parentFolder.listFiles(filter);
+		for (int j = 0; j < contents.length; j++) {
+			final File file = contents[j];
+			if (file.isDirectory()) {
+				if (!cleanFolder(file, filter, continueOnError, true) && !continueOnError)
+					return false;
+			} else {
+				if (!file.delete()) {
+					log.error("Couldn't delete " + file.getAbsolutePath());
+					if (!continueOnError)
+						return false;
+				}
 			}
 		}
+		if (deleteParentFolder) {
+			if (parentFolder.list().length == 0 && !parentFolder.delete()) {
+				log.error("Couldn't delete " + parentFolder.getAbsolutePath());
+				return false;
+			}
+		}
+		return true;
 	}
 
 }
